Message histories database operations moved into own file.
